+++ lib/draft_view/block/blocks/header_block.dart
@@ -1,4 +1,3 @@
-import 'dart:convert';
 import 'package:draft_view/draft_view/block/base_block.dart';
 import 'package:flutter/material.dart';