+++ lib/draft_view/block/blocks/image_block.dart
@@ -1,8 +1,10 @@
+import 'package:draft_view/draft_view/block/action_block.dart';
 import 'package:draft_view/draft_view/block/base_block.dart';
+import 'package:draft_view/draft_view/block/callbacks.dart';
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
 
-class ImageBlock extends BaseBlock {
+class ImageBlock extends ActionBlock {
   ImageBlock({
     required int depth,
     required int start,
@@ -12,6 +14,11 @@ class ImageBlock extends BaseBlock {
     required String text,
     required List<String> entityTypes,
     required String blockType,
+    ActionBuilder? actionBuilder,
+    OnTap? onTap,
+    OnDoubleTap? onDoubleTap,
+    OnLongPress? onLongPress,
+    List<BaseBlock>? children,
   }) : super(
           depth: depth,
           start: start,
@@ -21,6 +28,11 @@ class ImageBlock extends BaseBlock {
           text: text,
           entityTypes: entityTypes,
           blockType: blockType,
+          onTap: onTap,
+          onDoubleTap: onDoubleTap,
+          onLongPress: onLongPress,
+          actionBuilder: actionBuilder,
+          children: children,
         );
 
   ImageBlock copyWith({BaseBlock? block}) => ImageBlock(
@@ -32,24 +44,48 @@ class ImageBlock extends BaseBlock {
         data: block?.data ?? this.data,
         text: block?.text ?? this.text,
         blockType: block?.blockType ?? this.blockType,
+        actionBuilder: actionBuilder,
+        onTap: onTap,
+        onDoubleTap: onDoubleTap,
+        onLongPress: onLongPress,
       );
 
   @override
   InlineSpan render(BuildContext context, {List<InlineSpan>? children}) {
+    var actions = actionBuilder != null ? actionBuilder!(this) : null;
+
     return WidgetSpan(
       child: ImageComponent(
         url: data['src'],
         caption: data['description'],
+        onDoubleTap: onDoubleTap,
+        onTap: onTap,
+        onLongPress: onLongPress,
+        actions: actions,
+        imageBlock: this,
       ),
     );
   }
 }
 
 class ImageComponent extends StatefulWidget {
-  final String? url;
-  final String? caption;
+  final String url;
+  final String caption;
+  final OnLongPress? onLongPress;
+  final OnTap? onTap;
+  final OnDoubleTap? onDoubleTap;
+  final List<CupertinoContextMenuAction>? actions;
+  final ImageBlock imageBlock;
 
-  ImageComponent({this.url, this.caption});
+  ImageComponent({
+    required this.url,
+    required this.caption,
+    this.onTap,
+    this.onDoubleTap,
+    this.onLongPress,
+    this.actions,
+    required this.imageBlock,
+  });
 
   @override
   _ImageComponentState createState() => _ImageComponentState();
@@ -60,52 +96,74 @@ class _ImageComponentState extends State<ImageComponent> {
   Widget build(BuildContext context) {
     return Column(
       children: [
-        GestureDetector(
-          onTap: () {
-            showDialog(
-              context: context,
-              builder: (c) => ImageDetailView(
-                url: widget.url,
-                caption: widget.caption,
+        widget.url.isNotEmpty
+            ? Center(
+                child: _buildImage(),
+              )
+            : SizedBox.shrink(),
+        Hero(tag: Key(widget.caption), child: Text(widget.caption)),
+      ],
+    );
+  }
+
+  Widget _buildImage() {
+    var image = GestureDetector(
+      onDoubleTap: () => {
+        if (widget.onDoubleTap != null) widget.onDoubleTap!(widget.imageBlock)
+      },
+      onTap: () {
+        if (widget.onTap != null) {
+          widget.onTap!(widget.imageBlock);
+        } else {
+          if (widget.actions == null) {
+            Navigator.push(
+              context,
+              MaterialPageRoute(
+                builder: (c) =>
+                    ImageDetailView(url: widget.url, caption: widget.caption),
+              ),
+            );
+          }
+        }
+      },
+      child: Hero(
+        tag: widget.url,
+        child: Image.network(
+          widget.url,
+          fit: BoxFit.fitWidth,
+          loadingBuilder: (context, child, loadingProgress) {
+            if (loadingProgress == null) return child;
+            return Container(
+              height: 200,
+              color: Colors.grey.withOpacity(0.4),
+              child: Center(
+                child: CupertinoActivityIndicator(),
               ),
             );
           },
-          child: widget.url != null
-              ? Center(
-                  child: Image.network(
-                    widget.url!,
-                    fit: BoxFit.fitWidth,
-                    loadingBuilder: (context, child, loadingProgress) {
-                      if (loadingProgress == null) return child;
-                      return Container(
-                        height: 200,
-                        color: Colors.grey.withOpacity(0.4),
-                        child: Center(
-                          child: CupertinoActivityIndicator(),
-                        ),
-                      );
-                    },
-                  ),
-                )
-              : Container(
-                  height: 200,
-                  color: Colors.grey.withOpacity(0.4),
-                ),
         ),
-        Hero(
-          tag: Key("${widget.url}"),
-          child: Text("${widget.caption}"),
-        ),
-      ],
+      ),
     );
+    if (widget.actions == null) {
+      return image;
+    } else {
+      return CupertinoContextMenu(
+        child: image,
+        actions: widget.actions!,
+      );
+    }
   }
 }
 
 class ImageDetailView extends StatefulWidget {
-  final String? url;
-  final String? caption;
+  final String url;
+  final String caption;
 
-  const ImageDetailView({Key? key, this.url, this.caption}) : super(key: key);
+  const ImageDetailView({
+    Key? key,
+    required this.url,
+    required this.caption,
+  }) : super(key: key);
 
   @override
   _ImageDetailViewState createState() => _ImageDetailViewState();
@@ -114,8 +172,19 @@ class ImageDetailView extends StatefulWidget {
 class _ImageDetailViewState extends State<ImageDetailView> {
   final TransformationController _transformationController =
       TransformationController();
+  bool zoomed = false;
 
-  TapDownDetails? _doubleTapDetails;
+  late TapDownDetails _doubleTapDetails;
+
+  @override
+  void initState() {
+    super.initState();
+    _transformationController.addListener(() {
+      setState(() {
+        zoomed = _transformationController.value != Matrix4.identity();
+      });
+    });
+  }
 
   void _handleDoubleTapDown(TapDownDetails details) {
     _doubleTapDetails = details;
@@ -125,34 +194,50 @@ class _ImageDetailViewState extends State<ImageDetailView> {
     if (_transformationController.value != Matrix4.identity()) {
       _transformationController.value = Matrix4.identity();
     } else {
-      if (_doubleTapDetails != null) {
-        final position = _doubleTapDetails!.localPosition;
-        // For a 3x zoom
-        _transformationController.value = Matrix4.identity()
-          ..translate(-position.dx, -position.dy)
-          ..scale(2.0);
-      }
+      final position = _doubleTapDetails.localPosition;
+      // For a 3x zoom
+      _transformationController.value = Matrix4.identity()
+        ..translate(-position.dx, -position.dy)
+        ..scale(2.0);
     }
   }
 
   @override
   Widget build(BuildContext context) {
-    return Material(
-      child: Stack(
+    return Scaffold(
+      appBar: AppBar(
+        elevation: 0,
+        backgroundColor: Colors.transparent,
+        leading: IconButton(
+          color: Theme.of(context).buttonColor,
+          onPressed: () => Navigator.pop(context),
+          icon: Icon(
+            Icons.close,
+            color: Theme.of(context).iconTheme.color,
+          ),
+        ),
+      ),
+      body: Stack(
         children: [
-          if (widget.url != null)
-            Center(
-              child: GestureDetector(
-                onDoubleTapDown: _handleDoubleTapDown,
-                onDoubleTap: _handleDoubleTap,
+          if (widget.url.isNotEmpty)
+            GestureDetector(
+              onDoubleTapDown: _handleDoubleTapDown,
+              onDoubleTap: _handleDoubleTap,
+              child: Center(
                 child: InteractiveViewer(
                   transformationController: _transformationController,
-                  panEnabled: false,
-                  boundaryMargin: EdgeInsets.all(100),
+                  panEnabled: zoomed,
+                  boundaryMargin: EdgeInsets.all(30),
                   minScale: 0.5,
                   maxScale: 3,
-                  child: Image.network(
-                    widget.url!,
+                  child: Padding(
+                    padding: const EdgeInsets.only(bottom: 100),
+                    child: Hero(
+                      tag: widget.url,
+                      child: Image.network(
+                        widget.url,
+                      ),
+                    ),
                   ),
                 ),
               ),
@@ -163,23 +248,20 @@ class _ImageDetailViewState extends State<ImageDetailView> {
               width: MediaQuery.of(context).size.width,
               color: Colors.black,
               child: Padding(
-                padding: const EdgeInsets.all(8.0),
+                padding: const EdgeInsets.symmetric(
+                  horizontal: 20,
+                  vertical: 20,
+                ),
                 child: Hero(
-                  tag: Key("${widget.url}"),
+                  tag: widget.caption,
                   child: Text(
-                    "${widget.caption ?? ""}",
+                    widget.caption,
                     style: TextStyle(color: Colors.white, fontSize: 20),
                   ),
                 ),
               ),
             ),
           ),
-          Positioned(
-            child: IconButton(
-              onPressed: () => Navigator.pop(context),
-              icon: Icon(Icons.close),
-            ),
-          )
         ],
       ),
     );