+++ lib/draft_view/plugin/plugins/header_plugin.dart
@@ -1,4 +1,6 @@
-import '../../../draft_view.dart';
+import 'package:draft_view/draft_view/block/base_block.dart';
+import 'package:draft_view/draft_view/block/blocks/header_block.dart';
+import 'package:draft_view/draft_view/plugin/base_plugin.dart';
 
 class HeaderPlugin extends BasePlugin {
   @override