+++ test/block/image_block_test.dart
@@ -51,6 +51,7 @@ void main() {
       var component = MaterialApp(
         home: Scaffold(
           body: ImageDetailView(
+            url: '',
             caption: "abc.com",
           ),
         ),