+++ test/block/link_block_test.dart
@@ -1,11 +1,7 @@
-import 'dart:io' as io;
-
 import 'package:draft_view/draft_view.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:mockito/mockito.dart';
-
-import '../utils.dart';
 
 class MockBuildContext extends Mock implements BuildContext {}