+++ test/block/postsettings_block_test.dart
@@ -1,9 +1,7 @@
-import 'package:draft_view/draft_view.dart';
 import 'package:draft_view/draft_view/block/blocks/settings_block.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
 
-import '../test_data.dart';
 import 'blockquote_test.dart';
 
 void main() {