+++ test/converter/converter_test.dart
@@ -1,9 +1,5 @@
-import 'dart:math';
-
 import 'package:draft_view/draft_view.dart';
-import 'package:draft_view/draft_view/block/base_block.dart';
 import 'package:draft_view/draft_view/block/draft_object.dart';
-import 'package:draft_view/draft_view/converter/converter.dart';
 import 'package:flutter_test/flutter_test.dart';
 
 import '../test_data.dart';