+++ CHANGELOG.md
@@ -1,14 +1,22 @@
 <!-- @format -->
 
-## [0.9.0-nullsafety.0] - Basic functions implemented
+## [0.8.1] - Basic functions implemented
 
 - Implemented basic functions
 - A lot of tests
 - Null-safe!
 
+## [0.8.3] - Text Color support
+
+- Support color
+
+## [0.8.31] - Link Block
+
+- Support link block
+
 ## [0.9.1-nullsafety.0] - Audio Block
 
-- Support audio block (cannot play now. Wait until audioplayers has been migrated to null-safety)
+- Support audio block
 
 ## [0.9.2-nullsafety.0] - Link Block